Add replace option to trace_plot()

+++ project/utils.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
+import os
 
-
-def trace_plot(mc, path=None):
+def trace_plot(mc, path=None, replace=False):
     fig = plt.figure()
     ax1 = fig.add_subplot(111)
     ax1.set_title(f"Trace Plot for {len(mc)} iterations")
@@ -10,5 +10,6 @@ def trace_plot(mc, path=None):
     ax1.plot(range(len(mc)), mc)
     ax1.legend([f"beta {i}" for i in range(len(mc[0]))], loc="lower right")
     if path is not None:
-        fig.savefig(path)
+        if not os.path.exists(path) or replace:
+            fig.savefig(path)
     return fig
